+++ src/main/kotlin/totheking/traveljournals/configuration/WebSecurityConfig.kt
@@ -10,7 +10,7 @@ import org.springframework.security.crypto.factory.PasswordEncoderFactories
 import org.springframework.security.crypto.password.PasswordEncoder
 import org.springframework.security.web.SecurityFilterChain
 import totheking.traveljournals.auth.CustomUserDetailsService
-import totheking.traveljournals.user.application.UserCommandHandler
+import totheking.traveljournals.user.service.UserCommandHandler
 
 @EnableWebSecurity
 @Configuration
