+++ src/main/kotlin/totheking/traveljournals/travel/domain/TravelRepository.kt
@@ -1,4 +1,4 @@
-package totheking.traveljournals.travel.domain.repository
+package totheking.traveljournals.travel.domain
 
 import org.springframework.data.jpa.repository.JpaRepository
 import totheking.traveljournals.travel.domain.Travel
