+++ src/main/kotlin/totheking/traveljournals/user/domain/UserRepository.kt
@@ -1,0 +1,8 @@
+package totheking.traveljournals.user.domain
+
+import org.springframework.data.jpa.repository.JpaRepository
+import totheking.traveljournals.user.domain.User
+
+interface UserRepository : JpaRepository<User, Long> {
+    fun findByEmail(email: String): User?
+}
