--- src/main/kotlin/totheking/traveljournals/user/domain/repository/UserRepository.kt
@@ -1,8 +1,0 @@
-package totheking.traveljournals.user.domain.repository
-
-import org.springframework.data.jpa.repository.JpaRepository
-import totheking.traveljournals.user.domain.User
-
-interface UserRepository : JpaRepository<User, Long> {
-    fun findByEmail(email: String): User?
-}
